Reset changes.
