+++ packages/analytics-plugin-aws-pinpoint/package.json
@@ -1,14 +1,15 @@
 {
   "name": "@analytics/aws-pinpoint",
-  "version": "0.6.19",
+  "version": "0.6.20",
   "description": "AWS Pinpoint integration for 'analytics' module",
   "projectMeta": {
     "provider": {
       "name": "AWSPinpoint",
-      "url": "https://www.aws-pinpoint.com/"
+      "url": "https://aws.amazon.com/pinpoint/"
     },
     "platforms": {
-      "browser": "./src/browser.js"
+      "browser": "./src/browser.js",
+      "node": "./src/node.js"
     }
   },
   "keywords": [
@@ -42,18 +43,20 @@
     "lib",
     "README.md"
   ],
-  "homepage": "https://github.com/DavidWells/analytics#readme",
+  "homepage": "https://github.com/DavidWells/analytics/tree/master/packages/analytics-plugin-aws-pinpoint",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/DavidWells/analytics.git"
   },
   "dependencies": {
-    "@analytics/activity-utils": "^0.1.6",
+    "@analytics/activity-utils": "^0.1.7",
     "@analytics/localstorage-utils": "^0.1.5",
     "@analytics/queue-utils": "^0.1.1",
-    "@analytics/session-utils": "^0.1.10",
+    "@analytics/session-utils": "^0.1.11",
+    "@analytics/type-utils": "^0.3.1",
+    "@aws-sdk/client-pinpoint": "^3.31.0",
     "analytics-plugin-tab-events": "^0.2.0",
-    "analytics-utils": "^1.0.3",
+    "analytics-utils": "^1.0.4",
     "aws4fetch": "^1.0.13",
     "deepmerge": "^4.2.2"
   },
