+++ packages/analytics-plugin-aws-pinpoint/src/browser.js
@@ -1,6 +1,10 @@
-import { initialize, getStorageKey } from './pinpoint'
-import { CHANNEL_TYPES } from './pinpoint/constants'
-import * as PINPOINT_EVENTS from './pinpoint/events'
+import { initialize } from './pinpoint'
+import { getStorageKey } from './pinpoint/helpers/getStorageKey'
+import { CHANNEL_TYPES } from './pinpoint/helpers/constants'
+import * as PINPOINT_EVENTS from './pinpoint/helpers/events'
+import loadError from './utils/load-error'
+import formatEventData from './utils/format-event-data'
+import bootstrap from './utils/bootstrap'
 import { onUserActivity } from '@analytics/activity-utils'
 import { setItem, getItem, removeItem } from '@analytics/localstorage-utils'
 import {
@@ -9,7 +13,7 @@ import {
   extendSession,
   getTabSession,
   setTabSession,
-  getPageSession, 
+  getPageSession,
   setPageSession,
 } from '@analytics/session-utils'
 // import { onTabChange, isTabHidden } from 'analytics-plugin-tab-events'
@@ -25,7 +29,7 @@ const config = {
   // Pinpoint service region
   pinpointRegion: 'us-east-1',
   // Custom event mapping
-  eventMapping: {}
+  eventMapping: {},
 }
 
 /**
@@ -33,7 +37,7 @@ const config = {
  * @link https://docs.aws.amazon.com/pinpoint/latest/developerguide/
  * @param {object} pluginConfig - Plugin settings
  * @param {string} pluginConfig.pinpointAppId - AWS Pinpoint app Id for client side tracking
- * @param {function} pluginConfig.getCredentials - Async function to get AWS Cognito creds 
+ * @param {function} pluginConfig.getCredentials - Async function to get AWS Cognito creds
  * @param {string} [pluginConfig.pinpointRegion] - AWS Pinpoint region. Defaults to us-east-1
  * @param {string} [pluginConfig.appTitle] - The title of the app that's recording the event.
  * @param {string} [pluginConfig.appPackageName] - The name of the app package, such as com.example.my_app.
@@ -49,7 +53,7 @@ const config = {
  * })
  */
 function awsPinpointPlugin(pluginConfig = {}) {
-  let recordEvent 
+  let recordEvent
   let updateEndpoint
   let tabListener
   // let scrollDepthMax = 0
@@ -57,7 +61,7 @@ function awsPinpointPlugin(pluginConfig = {}) {
 
   /* Page-session (on route changes) */
   let hasPageFiredOnce = false
-  
+
   function stopSession() {
     const currentSessionData = getSession()
     if (pluginConfig.debug) {
@@ -87,23 +91,12 @@ function awsPinpointPlugin(pluginConfig = {}) {
     name: 'aws-pinpoint',
     config: {
       ...config,
-      ...pluginConfig
+      ...pluginConfig,
     },
-    bootstrap: (pluginApi) => {
-      const { config, instance } = pluginApi
-      /* Load aws-pinpoint script after userId exists */
-      if (config.disableAnonymousTraffic && !instance.user('userId')) {
-        instance.once('identifyStart', ({ plugins }) => {
-          const self = plugins['aws-pinpoint']
-          if (!self.loaded()) {
-            instance.loadPlugin('aws-pinpoint')
-          }
-        })
-      }
-    },
+    bootstrap: bootstrap,
     initialize: ({ config, instance }) => {
       const { disableAnonymousTraffic, debug } = config
-      const logger = (debug) ? console.log : () => {}
+      const logger = debug ? console.log : () => {}
       /* Disable pinpoint if user is not yet identified. */
       const state = instance.getState()
       const userDetails = state.user || {}
@@ -118,9 +111,13 @@ function awsPinpointPlugin(pluginConfig = {}) {
       logger('initPageSession', initPageSession)
       logger('initTabSession', initTabSession)
       logger('initSessionData', initSessionData)
-  
+
       /* If anonId has changed, refresh session details */
-      if (initSessionData && initSessionData.anonId && initSessionData.anonId !== anonymousId) {
+      if (
+        initSessionData &&
+        initSessionData.anonId &&
+        initSessionData.anonId !== anonymousId
+      ) {
         logger('anonId different refresh session details')
         // console.log('anonId', anonymousId)
         // console.log('initSessionData.anonId', initSessionData.anonId)
@@ -185,7 +182,7 @@ function awsPinpointPlugin(pluginConfig = {}) {
             title: document.title,
             host: window.location.hostname,
             url: window.location.origin + window.location.pathname,
-            ...utmParams
+            ...utmParams,
           }
         },
         // Custom event mapping
@@ -296,16 +293,12 @@ function awsPinpointPlugin(pluginConfig = {}) {
       }
       const endpoint = {}
       const userInfo = {}
-  
+
       if (userId) {
         userInfo.UserId = userId
       }
       if (traits && Object.keys(traits).length) {
         userInfo.UserAttributes = traits
-      }
-      if (traits.email) {
-        endpoint.Address = traits.email,
-        endpoint.ChannelType = CHANNEL_TYPES.EMAIL
       }
       if (Object.keys(userInfo).length) {
         endpoint.User = userInfo
@@ -321,26 +314,6 @@ function awsPinpointPlugin(pluginConfig = {}) {
     },
     loaded: () => !!recordEvent,
   }
-}
-
-function loadError() {
-  throw new Error('Pinpoint not loaded')
-}
-
-function formatEventData(obj) {
-  return Object.keys(obj).reduce((acc, key) => {
-    const value = obj[key]
-    if (typeof value === 'number') {
-      acc.metrics[key] = value
-    }
-    if (typeof value === 'string' || typeof value === 'boolean') {
-      acc.attributes[key] = value
-    }
-    return acc
-  }, {
-    attributes: {},
-    metrics: {}
-  })
 }
 
 function getTabs() {
@@ -360,9 +333,9 @@ function tabLoadEventHandler() {
   setTabs(tabs)
 }
 function tabUnloadEventHandler() {
-  let hash= sessionStorage.getItem('TabHash')
+  let hash = sessionStorage.getItem('TabHash')
   let tabs = getTabs()
-  delete tabs[hash];
+  delete tabs[hash]
   setTabs(tabs)
 }
