+++ packages/analytics-plugin-aws-pinpoint/src/pinpoint/helpers/get-event-name.js
@@ -1,4 +1,3 @@
-
 export default function getEventName(key, eventMap = {}) {
   return eventMap[key] || key
 }
