+++ packages/analytics-plugin-aws-pinpoint/src/pinpoint/helpers/getStorageKey.js
@@ -1,0 +1,5 @@
+export const ENDPOINT_KEY = '__endpoint'
+
+export function getStorageKey(id) {
+  return `${ENDPOINT_KEY}.${id}`
+}
