+++ packages/analytics-plugin-aws-pinpoint/src/utils/bootstrap.js
@@ -1,0 +1,12 @@
+export default function bootstrap(pluginApi) {
+  const { config, instance } = pluginApi
+  /* Load aws-pinpoint script after userId exists */
+  if (config?.disableAnonymousTraffic && !instance.user('userId')) {
+    instance.once('identifyStart', ({ plugins }) => {
+      const self = plugins['aws-pinpoint']
+      if (!self.loaded()) {
+        instance.loadPlugin('aws-pinpoint')
+      }
+    })
+  } 
+}
