--- packages/analytics-plugin-aws-pinpoint/src/utils/in-browser.js
@@ -1,2 +1,0 @@
-const inBrowser = typeof window !== 'undefined'
-export default inBrowser
