+++ packages/analytics-plugin-aws-pinpoint/src/utils/load-error.js
@@ -1,0 +1,3 @@
+export default function loadError() {
+  throw new Error('Pinpoint not loaded')
+}
